"Rename Admin side tokens"
